Fix foceus shift on search
This change improves a few things about how dialogs are handled. Most importantly it changes the behaviour when the search or filters change to prevent the focus shifting. Note this is achieved via a currious side-effect so this could be considered a hack. The changes:

* reduce number of state elements in Home as there were redundencies
* fix up effects in Dialog so were only calling show/close if we need to
* Don't render the dialog if we don't have to. This is a hack to prevent the focus change behaviour
* only deselect the selectedFeature on search change if the new list does not contain the selected feature

Basically the issue occurs because the dialog returns focus to the previous selected input that triggered it's opening on close. That is the button list of cards. There doesn't seem to be a way of preventing this without moving away from html dialog in the component library